@@ -1,32 +1,30 @@
 name: ishangavidusha
 description: Ishanga Vidusha's Portfolio Website.
-publish_to: 'none' # Remove this line if you wish to publish to pub.dev
+publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
   sdk: ">=2.16.1 <3.0.0"
 dependencies:
+  animated_text_kit: ^4.2.2
+  collection: ^1.16.0
   flutter:
     sdk: flutter
-
-  provider: ^6.0.2
-  go_router: ^3.0.1
-  url_launcher: ^6.0.20
-  shared_preferences: ^2.0.13
-  intl: ^0.17.0
-  google_fonts: ^2.3.1
-  font_awesome_flutter: ^9.2.0
-  animated_text_kit: ^4.2.1
-  collection: ^1.15.0
+  font_awesome_flutter: ^10.3.0
+  go_router: ^5.2.4
+  google_fonts: ^3.0.1
+  intl: ^0.18.0
+  provider: ^6.0.5
+  shared_preferences: ^2.0.15
+  url_launcher: ^6.1.7
 
 dev_dependencies:
+  flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
-  flutter_lints: ^1.0.0
 
 flutter:
   uses-material-design: true
-
   # To add assets to your application, add an assets section, like this:
   # assets:
   #   - images/a_dot_burr.jpeg
