@@ -1,16 +1,16 @@
 import '../utils/enum_to_string.dart';
 
-enum APP_PAGES {
+enum Pages {
   splash,
   home,
 }
 
-extension AppPageExtension on APP_PAGES {
+extension AppPageExtension on Pages {
   String toPath({bool subRoute = false}) {
-    if (this == APP_PAGES.home) {
+    if (this == Pages.home) {
       return subRoute ? "" : "/";
     }
-    return subRoute ? EnumToString.convertToString(this).toLowerCase() : "/" + EnumToString.convertToString(this).toLowerCase();
+    return subRoute ? EnumToString.convertToString(this).toLowerCase() : "/${EnumToString.convertToString(this).toLowerCase()}";
   }
 
   String toPathName() {
